EPMLABSBRN-46 POST end-point: Write a result of finishing exercise to User history

+++ src/main/kotlin/com/epam/brn/model/UserDetails.kt
@@ -1,18 +1,18 @@
 package com.epam.brn.model
 
+import javax.persistence.Column
 import javax.persistence.Entity
-import javax.persistence.Table
-import javax.persistence.Id
 import javax.persistence.GeneratedValue
 import javax.persistence.GenerationType
-import javax.persistence.Column
+import javax.persistence.Id
+import javax.persistence.Table
 
 @Entity
 @Table(name = "USER_DETAILS")
 data class UserDetails @JvmOverloads constructor(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Int,
+    val id: Long,
     @Column(nullable = false)
     val name: String,
     @Column(nullable = true)
